% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFilters.R
\name{allFilters}
\alias{allFilters}
\title{allFilters Function}
\usage{
allFilters(Estado, year, typology)
}
\arguments{
\item{Estado}{Enter the States you want.}

\item{year}{Enter the years you want.}

\item{typology}{Enter the criminal types you want.}
}
\description{
The allFilters function allows simultaneous filtering through the UF,
year, and criminal typology.
}
\examples{
acre_15_homc <- allFilters("Acre",2015,"Homicídio doloso")

acre_sp_2016_18_estp_homic <- allFilters(c("Acre","São Paulo"),
2016:2018,c("Estupro","Homicídio doloso"))



}
