% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimeType.R
\name{crimeType}
\alias{crimeType}
\title{crimeType Function}
\usage{
crimeType(typology)
}
\arguments{
\item{typology}{Enter the criminal types you want.}
}
\description{
The crimeType function filters the data based on the criminal typology provided.
}
\examples{
estupro <- crimeType("Estupro")

estu_e_homic <- crimeType(c("Estupro","Homicídio doloso"))



}
