% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sinespvde_data.R
\name{get_sinespvde_data}
\alias{get_sinespvde_data}
\title{get_sinesp-vde_data Function}
\usage{
get_sinespvde_data(
  state = "all",
  city = "all",
  category = "all",
  typology = "all",
  year = "all",
  granularity = "month",
  pivot = F,
  geom = F,
  relative_values = F
)
}
\description{
This function is responsible for collecting criminal data directly from the SINESP database.
The data collected by this function starts temporally in January 2015 and goes until December 2022,
covering crimes such as: Estupro (Rape), Furto de veículo (Vehicle Theft), Homicídio doloso (Intentional Homicide),
Lesão corporal seguida de morte (Bodily Injury Followed by Death), Roubo a instituição financeira (Robbery of Financial Institution),
Roubo de carga (Cargo Theft), Roubo de veículo (Vehicle Robbery), Roubo seguido de morte (Latrocínio) (Robbery Followed by Death)
and Tentativa de homicídio (Attempted Homicide).
}
\examples{
dados <- get_sinespvde_data()


}
