% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{data}
\name{getData}
\alias{getData}
\title{getData Function}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2592 rows and 12 columns.
}
\usage{
getData
}
\description{
This function is responsible for collecting criminal data directly from the SINESP database.
The data collected by this function starts temporally in January 2015 and goes until December 2022,
covering crimes such as "Rape", "Vehicle Theft", "Intentional Homicide", "Bodily Injury Followed by Death",
"Robbery of Financial Institution", "Cargo Theft", "Vehicle Robbery", "Robbery Followed by Death"
and "Attempted Homicide".
}
\keyword{datasets}
