% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFData.R
\name{UFData}
\alias{UFData}
\title{UFData Function}
\usage{
UFData(Estado)
}
\arguments{
\item{Estado}{Use the UF's name to select the desired States.}
}
\description{
The UFData function filters the data based on the name of the Brazilian
Federative Unit provided.
}
\examples{
SP = UFData("São Paulo")

Sudeste <- UFData(c("São Paulo","Minas Gerais","Rio de Janeiro","Espírito Santo"))


}
