% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearData.R
\name{yearData}
\alias{yearData}
\title{yearData Function}
\usage{
yearData(year)
}
\arguments{
\item{year}{Give years to be selected.}
}
\description{
The yearData function filters the data based on the years provided.
}
\examples{
y2016 = yearData(2016)

y2016_18 <- yearData(2016:2018)


}
