% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFData.R
\name{UFData}
\alias{UFData}
\title{UFData Function}
\usage{
UFData(Estado)
}
\arguments{
\item{Estado}{Use the UF's name to select the one you want.}
}
\description{
The UFData function filters the data based on the name of the Brazilian
Federative Unit provided.
}
